+++ NEWS.md
@@ -2,8 +2,15 @@
 
 * feat: All learners now support the `predict_raw` flag to store raw prediction objects from upstream models.
 * fix: using offset during prediction for `xgboost` learners
-* fix: importance in doc examples is now printed correctly 
+* fix: importance in doc examples is now printed correctly
 * feat: add parameter `local.importance` to ranger learners
+* feat: `cv_glmnet` learners gained a `seed` parameter to make the random fold assignment of `cv.glmnet`, and therefore the selected lambda, reproducible.
+* fix: compatibility with `glmnet` 5.0
+* fix: use of `relax = TRUE` with glmnet learners
+* refactor: went through `glmnet` parameters and updated their order, defaults, removed those that are not utilized by corresponding learners, etc.
+* fix: using offset during prediction for `xgboost` learners.
+* fix: importance in doc examples is now printed correctly.
+* feat: add parameter `local.importance` to ranger learners.
 
 # mlr3learners 0.14.0